pyproject.toml
[project]
name = "hebergement-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayoub Mouchrif",email = "ayoub.mouchrif-ext@um6p.ma"}
]
packages = [{ include = "app" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "databases (>=0.9.0,<0.10.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "dynaconf (>=3.2.10,<4.0.0)",
    "json-log-formatter (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)",
    "common @ git+https://github.com/MouchrifAyoub/common.git",
    "httpx (>=0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
pytest = "^8.3.5"
httpx = "0.26.0"
uvicorn = "^0.34.0"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.26.0"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"